Fix playbook terminate instance
Fixes a bug where trying to call the playbook from
popen never returns.  Issue was we were actually
running the novaclient call on the instance, so we
in essence killed ourself and caused ansible to be
unable to complete, leaving the process hung.

Just needed to modify the task to use localhost instead
of devstack_instance.

Number of other little cleanups, including a working
testit.py that just runs the executor code for testing,
that way we don't need to wait on a valid change from the
gerrit stream, and of course we cut out the voting piece.