Add support for multiple nics on instance
Previously we worked on the assumption that you only
had a single private IF to the instance.

This makes some slight mods to the Ansible tasks to enable
us to parse things out when adding a second IF.

Also adds a new task to enable the IF on the instance after
creation.