Finishing touches to make it work
Move the teardown to it's own independent task, otherwise
if a preceeding task fails, ansible bales out and you never
delete the instance.  Would be cool if there were a "cleanup"
clause in Ansible but doesn't seem to exist at the moment.

Also fixed up the submit to gerrit string to use the regex
catcher to make it look "pretty".

* NOTE
Been running a week or so now and could probably turn on voting
but just having it report on SUCCESS is all I'm really intrested
in right now, so good enough.

** NOTE
Starting to look at how to work the ansible approach in to the
format OpenStack Infra code base.