Update to publish failures
This tool was initially only reporting and publishing on
success.  This changeset includes some signficiant changes
to the ansible modules and to the executor to ensure we
publish all the logs we can in the case of a failure.

Also improve the logging strcture to create a seperate
log file for the ansible tasks themeselve.  This way we
can tell "why" a run failed if it had to do with deployment.