Remove hard coded items and revert db feature
Seems there are folks out there using this for their own
CI deployment which is AWESOME!!  There's lots of room for
imporovement here so don't be shy with your pull requests.

At the same time, there seems to be a couple of folks that
aren't really "looking" at the code and submitting comments
to gerrit using my CI name (my fault since it's hardcoded in there).

This patch removes all of my system specific things like names and
IP's.  It also modifies the sample conf file to have pure
nonsense entries so one will be forced to update.

In addition, my db integration seems to have some issues, for now
I'm just reverting that and I'll come back to it at some point
in the future as a nice to have.