Add support for recheck
This just adds an additional filter on events to detect
recheck requests.

This breaks the filter method in two to keep it from being a gnarly
mess of if/else/if.  To make this work we have a filter that checks
the event for the comment "sos-ci.conf.recheck_string" where recheck
string is a new value added to the sos-ci.conf file and is for
example 'recheck solidfire'.

We also check and make sure we're on master for the recheck as well.