Add results.html geneartion and reporting
One of the requirements that was msising from sos-ci is the
generatio and publishing of results.html.

This patch steals the results2html.py script and adds it to the
sos/scripts, and adds a step to generate the subunit results and
html file and publish them with the logs bundle.