Fix up issues with log collection
The subunit2html file was getting sent over without
execute permissions, just add an explicit task to
chmod it so we won't have issues here.

Also, fix up recheck, make sure that if we already have
a dir for that run locally we nuke it and create a fresh
one (the data should be published anyway).