Use the subunit2sql python api directly
Following the instructions outlined on [1] this commit switches to
using subunit2sql's python api directly to store result streams in
the DB instead of subprocessing out to call the subunit2sql scripts.
As part of this some extra metadata bits are also added to the DB to
add some extra info to track.

[1] http://docs.openstack.org/developer/subunit2sql/api.html#example-usage-patterns