Handle gerrit login issues
Lateley we've been having a number of maintenance windows for
Gerrit.  Currently the os_ci code just pukes and exits if/when
it can't reach Gerrit, and I have to remember to go back and
restart.

This patch just adds an infinite retry loop around the Gerrit
connect routine, this way nobody has to "care" about this and
we'll just pick up the connection when Gerrit is back online.