Auto-Reconnect to gerrit
This patch removes the annoying logging that the queue is empty,
and more importantly adds a check for the gerrit connection and
reconnects for those times when gerrit goes offline for maintenance
or whatver.  Otherwise I have to notice we quit getting things in
the queue and remember to restart, or catch the email on dev list
that gerrit is going down etc.