Create a branch specifically for SF
This is just a slightly modified version specifically for
my own deployment.  This isn't used a ton by other folks
so we never did things like make all of the vars generic
and such.  Well, at least nobody that forked it ever had
and interested in working on it and submitting PR's.

So, I'll just keep an independent branch with all my own
tweaks and such for now.