Update a bunch of stuff
This is bad form, but things have diverged drastically between the
changes upstream and some hardening to the tool while running over
the years.

This patch adds a bunch of changes and gets things back up to speed.

The most important thing here is the addition of an explicit "close"
call to the paramiko connection used to submit votes/comments to
gerrit.  Turns out that by leaving those to GC they were piling up on
the gerrit server and we had hundreds of hanging connections on the
remote server.